Bug fix to GSAlign processes

+++ modules/processes/GSAlign.nf
@@ -14,10 +14,19 @@ process gsalign_same{
   
     script:
     """
-    GSAlign -r ${source} -q ${tgtfile} -sen -t ${task.cpus} -idy 98 -o ${srcname}.${tgtname}.tmp
-    maf-convert psl ${srcname}.${tgtname}.tmp.maf |
-        liftUp -type=.psl stdout $srclift warn stdin |
-        liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin && rm ${srcname}.${tgtname}.tmp.*
+    bwt_index ${srcfile} ${srcname} && chmod a+rw ${srcname}.*
+    GSAlign -i ${srcname} -q ${tgtfile} -sen -t ${task.cpus} -idy 90 -no_vcf -o ${srcname}.${tgtname}.tmp
+    if [ -e ${srcname}.${tgtname}.tmp.maf ]; then
+        sed 's/ref.//g' ${srcname}.${tgtname}.tmp.maf | 
+            sed 's/qry.//g' |
+            maf-convert psl - |
+            liftUp -type=.psl stdout $srclift warn stdin |
+            liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin && rm ${srcname}.${tgtname}.tmp.*
+    else
+        echo "##maf version=1 " | 
+            liftUp -type=.psl stdout $srclift warn stdin |
+            liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin 
+    fi
     """
 }
 
@@ -35,10 +44,19 @@ process gsalign_near{
   
     script:
     """
-    GSAlign -r ${source} -q ${tgtfile} -sen -t ${task.cpus} -idy 95 -o ${srcname}.${tgtname}.tmp
-    maf-convert psl ${srcname}.${tgtname}.tmp.maf |
-        liftUp -type=.psl stdout $srclift warn stdin |
-        liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin && rm ${srcname}.${tgtname}.tmp.*
+    bwt_index ${srcfile} ${srcname} && chmod a+rw ${srcname}.*
+    GSAlign -i ${srcname} -q ${tgtfile} -sen -t ${task.cpus} -idy 80 -no_vcf -o ${srcname}.${tgtname}.tmp
+    if [ -e ${srcname}.${tgtname}.tmp.maf ]; then
+        sed 's/ref.//g' ${srcname}.${tgtname}.tmp.maf | 
+            sed 's/qry.//g' |
+            maf-convert psl - |
+            liftUp -type=.psl stdout $srclift warn stdin |
+            liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin && rm ${srcname}.${tgtname}.tmp.*
+    else
+        echo "##maf version=1 " | 
+            liftUp -type=.psl stdout $srclift warn stdin |
+            liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin 
+    fi
     """
 }
 
@@ -56,10 +74,19 @@ process gsalign_medium{
   
     script:
     """
-    GSAlign -r ${source} -q ${tgtfile} -t ${task.cpus} -idy 90 -o ${srcname}.${tgtname}.tmp
-    maf-convert psl ${srcname}.${tgtname}.tmp.maf |
-        liftUp -type=.psl stdout $srclift warn stdin |
-        liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin && rm ${srcname}.${tgtname}.tmp.*
+    bwt_index ${srcfile} ${srcname} && chmod a+rw ${srcname}.*
+    GSAlign -i ${srcname} -q ${tgtfile} -t ${task.cpus} -idy 75 -no_vcf -o ${srcname}.${tgtname}.tmp
+    if [ -e ${srcname}.${tgtname}.tmp.maf ]; then
+        sed 's/ref.//g' ${srcname}.${tgtname}.tmp.maf | 
+            sed 's/qry.//g' |
+            maf-convert psl - |
+            liftUp -type=.psl stdout $srclift warn stdin |
+            liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin && rm ${srcname}.${tgtname}.tmp.*
+    else
+        echo "##maf version=1 " | 
+            liftUp -type=.psl stdout $srclift warn stdin |
+            liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin 
+    fi
     """
 }
 
@@ -77,10 +104,19 @@ process gsalign_far{
   
     script:
     """
-    GSAlign -r ${source} -q ${tgtfile} -t ${task.cpus} -idy 80 -o ${srcname}.${tgtname}.tmp
-    maf-convert psl ${srcname}.${tgtname}.tmp.maf |
-        liftUp -type=.psl stdout $srclift warn stdin |
-        liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin && rm ${srcname}.${tgtname}.tmp.*
+    bwt_index ${srcfile} ${srcname} && chmod a+rw ${srcname}.*
+    GSAlign -i ${srcname} -q ${tgtfile} -t ${task.cpus} -idy 70 -no_vcf -o ${srcname}.${tgtname}.tmp
+    if [ -e ${srcname}.${tgtname}.tmp.maf ]; then
+        sed 's/ref.//g' ${srcname}.${tgtname}.tmp.maf | 
+            sed 's/qry.//g' |
+            maf-convert psl - |
+            liftUp -type=.psl stdout $srclift warn stdin |
+            liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin && rm ${srcname}.${tgtname}.tmp.*
+    else
+        echo "##maf version=1 " | 
+            liftUp -type=.psl stdout $srclift warn stdin |
+            liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin
+    fi
     """
 }
 
@@ -98,9 +134,18 @@ process gsalign_custom{
   
     script:
     """
-    GSAlign -r ${source} -q ${tgtfile} -t ${task.cpus} ${params.custom} -o ${srcname}.${tgtname}.tmp
-    maf-convert psl ${srcname}.${tgtname}.tmp.maf |
-        liftUp -type=.psl stdout $srclift warn stdin |
-        liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin && rm ${srcname}.${tgtname}.tmp.*
+    bwt_index ${srcfile} ${srcname} && chmod a+rw ${srcname}.*
+    GSAlign -i ${srcname} -q ${tgtfile} -t ${task.cpus} ${params.custom} -no_vcf -o ${srcname}.${tgtname}.tmp
+    if [ -e ${srcname}.${tgtname}.tmp.maf ]; then
+        sed 's/ref.//g' ${srcname}.${tgtname}.tmp.maf | 
+            sed 's/qry.//g' |
+            maf-convert psl - |
+            liftUp -type=.psl stdout $srclift warn stdin |
+            liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin && rm ${srcname}.${tgtname}.tmp.*
+    else
+        echo "##maf version=1 " | 
+            liftUp -type=.psl stdout $srclift warn stdin |
+            liftUp -type=.psl -pslQ ${srcname}.${tgtname}.psl $tgtlift warn stdin
+    fi
     """
 }
